fix(telegram): surface rate-limit backoffs and retry sendOrEdit 429s (#941)
Two remaining causes of "session appears to hang / times out" on Telegram
when subagents are in flight:

FM-1 — sendOrEdit silently swallowed ALL non-400 Telegram errors, including
429 flood-control during rapid in-turn edits from subagent tool calls. The
preview message froze while the turn kept running. Now retries 429s once
using the server's retry_after via the existing floodRetryAfterMs helper.

FM-4 — The provider emits rate_limit events during Claude API backoffs
(429/529), but the streaming loop had no handler. The message sat frozen
for multi-minute retries; if the backoff exceeded the 180s watchdog, a
false "Response timed out" fired — killing a turn that was making progress.
Now renders a visible status line and the watchdog resets on each event.

Extraction: activity label formatting (humanizeToolActivity,
formatTelegramActivity, formatTelegramAgentLabel, renderSubagentFooter)
moved to streaming.activity.ts to make room under the file-size ceiling.
Re-exported from streaming.ts so all existing imports are unchanged.
